# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auths-controller"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "LoginRequest",
                        "name": "LoginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.LoginResponse"
                        }
                    }
                }
            }
        },
...
        "/history-productions": {
            "get": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Fetch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start date",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end date",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            },
            "post": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Store",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "HistoryProduction",
                        "name": "HistoryProduction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.HistoryProduction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            }
        },
        "/history-productions/barchart": {
            "get": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Barchart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            }
        },
        "/history-productions/stat": {
            "get": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Stat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductionStats"
                        }
                    }
                }
            }
        },
        "/history-productions/{id}": {
            "get": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "FindById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "historyProduction id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            },
            "put": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "UpdateHistoryProduction",
                        "name": "UpdateHistoryProduction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateHistoryProduction"
                        }
                    },
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            },
            "delete": {
                "description": "Put all mandatory parameter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history-productions-controller"
                ],
                "summary": "Destroy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "historyProduction id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HistoryProductions"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.HistoryProduction": {
            "type": "object",
            "properties": {
                "market_demand": {
                    "type": "integer"
                },
                "period_date": {
                    "type": "string"
                },
                "production": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "request.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "request.UpdateHistoryProduction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "market_demand": {
                    "type": "integer"
                },
                "period_date": {
                    "type": "string"
                },
                "production": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "response.HistoryProductionStats": {
            "type": "object",
            "properties": {
                "max_market_demand": {
                    "type": "integer"
                },
                "max_production": {
                    "type": "integer"
                },
                "max_stock": {
                    "type": "integer"
                },
                "min_market_demand": {
                    "type": "integer"
                },
                "min_production": {
                    "type": "integer"
                },
                "min_stock": {
                    "type": "integer"
                }
            }
        },
        "response.HistoryProductions": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "market_demand": {
                    "type": "integer"
                },
                "period_date": {
                    "type": "string"
                },
                "production": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "response.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}